remove MemorialState from AdminState, made id separate. Also add memorial module on server side
