+++ client/src/constants/actions.ts
@@ -14,15 +14,18 @@ export const PUBLICATION_AT = {
 };
 
 export const MARKERS_AT = {
-  ADD_MEMORIAL_MARKER: 'ADD_MEMORIAL_MARKER',
   ADD_MEMORIAL_TYPE: 'ADD_MEMORIAL_TYPE',
   EDIT_MEMORIAL_TYPE: 'EDIT_MEMORIAL_TYPE',
   DELETE_MEMORIAL_TYPE: 'DELETE_MEMORIAL_TYPE',
-  EDIT_MEMORIAL: 'EDIT_MEMORIAL',
-  DELETE_MEMORIAL_MARKER: 'DELETE_MEMORIAL_MARKER',
-  GET_ALL_MEMORIALS: 'GET_ALL_MEMORIALS',
   GET_ALL_MEMORIAL_TYPES: 'GET_ALL_MEMORIAL_TYPES',
+};
+
+export const MEMORIAL_AT = {
+  ADD_MEMORIAL_MARKER: 'ADD_MEMORIAL_MARKER',
+  EDIT_MEMORIAL: 'EDIT_MEMORIAL',
+  GET_ALL_MEMORIALS: 'GET_ALL_MEMORIALS',
   GET_CURRENT_MEMORIAL: 'GET_CURRENT_MEMORIAL',
+  DELETE_MEMORIAL_MARKER: 'DELETE_MEMORIAL_MARKER',
 };
 
 export const CITY_MARKERS_AT = {
