+++ client/src/pages/AdminMemorialsPage/MemorialContextMenu/MemorialContextMenu.tsx
@@ -5,11 +5,17 @@ import { useTranslation } from 'react-i18next';
 import { deleteMemorialMarkerRequest } from '@src/store/adminMarkers/action';
 import { IMemorialMarkerContextMenuProps } from '@src/pages/AdminMemorialsPage/MemorialContextMenu/types';
 import { useNavigate } from 'react-router-dom';
+import { Loader } from '@src/components/Loader/Loader';
+import { useAppSelector } from '@src/hooks/hooks';
+import { selectMemorialMarkersLoadingItems } from '@src/store/memorialMarkers/selectors';
 import * as S from './style';
+import { MemorialLoaderWrapper } from './style';
 
 const MemorialContextMenu: React.FC<IMemorialMarkerContextMenuProps> = ({ marker }) => {
   const [isDeleteOpen, setDeleteOpen] = useState(false);
+  const isLoadingMemorials = useAppSelector(selectMemorialMarkersLoadingItems);
   const navigate = useNavigate();
+  const { _id } = marker;
   const { t } = useTranslation();
   
   const toggleDeleteModalHandler = () => {
@@ -17,7 +23,7 @@ const MemorialContextMenu: React.FC<IMemorialMarkerContextMenuProps> = ({ marker
   };
   
   const toggleEditHandler = () => {
-    navigate(`/mapmem-admin/memorials/edit/${marker._id}`);
+    navigate(`/mapmem-admin/memorials/edit/${_id}`);
   };
   
   return (
@@ -25,16 +31,17 @@ const MemorialContextMenu: React.FC<IMemorialMarkerContextMenuProps> = ({ marker
       {isDeleteOpen && (
         <Modal outsideHandler={toggleDeleteModalHandler}>
           <DeleteModal
-            itemId={marker._id as string}
+            itemId={_id as string}
             action={deleteMemorialMarkerRequest}
             text={t('deleteMemorialMarkerText')}
             onClose={toggleDeleteModalHandler}
           />
         </Modal>
       )}
+      {isLoadingMemorials.includes(_id) && <S.MemorialLoaderWrapper><Loader size={15}/></S.MemorialLoaderWrapper>}
       <S.MemorialContextMenuWrapper>
-        <S.MemorialContextItem onClick={toggleEditHandler}>Редагувати</S.MemorialContextItem>
-        <S.MemorialContextItem onClick={toggleDeleteModalHandler}>Видалити</S.MemorialContextItem>
+        <S.MemorialContextItem onClick={toggleEditHandler}>{t('edit')}</S.MemorialContextItem>
+        <S.MemorialContextItem onClick={toggleDeleteModalHandler}>{t('delete')}</S.MemorialContextItem>
       </S.MemorialContextMenuWrapper>
     </>
   );
