+++ client/src/pages/AdminMemorialsPage/MemorialContextMenu/style.ts
@@ -27,3 +27,12 @@ export const MemorialContextItem = styled.li`
     }
   }
 `;
+
+export const MemorialLoaderWrapper = styled.div`
+  background-color: rgba(0, 0, 0, 0.06);
+  position: absolute;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+`;
