--- client/src/store/adminMarkers/types.ts
@@ -1,11 +1,0 @@
-import { IAdminCityMarker, IAdminMemorialMarker, IAdminMemorialType, IMemorialMarker } from '@src/types/markers.types';
-
-export interface IAdminMarkersState {
-  isLoading: boolean,
-  cityMarkers: IAdminCityMarker[],
-  memorialMarkers: IAdminMemorialMarker[],
-  loadingItem: string[],
-  memorialTypes: IAdminMemorialType[],
-  currentMemorial: IMemorialMarker | null,
-  isAddEditCompleted: boolean,
-}
