+++ client/src/store/markerTypes/types.ts
@@ -1,0 +1,7 @@
+import { IAdminMemorialType } from '@src/types/markers.types';
+
+export interface IAdminMarkerTypesState {
+  isLoading: boolean,
+  data: IAdminMemorialType[],
+  loadingItem: string[],
+}
