+++ server/src/components/map/map.controller.ts
@@ -6,15 +6,11 @@ import {
   Param,
   Post,
   Query,
-  UploadedFiles,
   UseGuards,
-  UseInterceptors,
 } from '@nestjs/common';
 import { MapService } from './map.service';
 import { JwtAuthGuard } from '../../guards/jwtAuth.guard';
-import { FileFieldsInterceptor } from '@nestjs/platform-express';
 import { FormDataRequest } from 'nestjs-form-data';
-import { CreateMemorialDto } from '../../dto/createMemorial.dto';
 import { CreateMemorialTypeDto } from '../../dto/createMemorialType.dto';
 import { ROUTES } from '../../constants/routes';
 @Controller(ROUTES.MAP.DEFAULT)
@@ -25,11 +21,6 @@ export class MapController {
   getInfo(@Query() query: Record<string, string>) {
     const filters = query.filters.split('_%_').filter((item) => item);
     return this.mapService.getInfo(filters);
-  }
-
-  @Get(ROUTES.MAP.GET_ALL_MEMORIALS)
-  getAllMemorials() {
-    return this.mapService.getAllMemorials();
   }
 
   @Get(ROUTES.MAP.GET_ALL_MEMORIAL_TYPES)
@@ -55,37 +46,5 @@ export class MapController {
   @UseGuards(JwtAuthGuard)
   deleteMemorialType(@Param('id') id) {
     return this.mapService.deleteMemorialType(id);
-  }
-
-  @Get(ROUTES.MAP.GET_MEMORIAL)
-  getMemorial(@Param('id') id) {
-    return this.mapService.getMemorial(id);
-  }
-
-  @Delete(ROUTES.MAP.DELETE_MEMORIAL_MARKER)
-  @UseGuards(JwtAuthGuard)
-  deleteMemorial(@Param('id') id) {
-    return this.mapService.deleteMemorial(id);
-  }
-
-  @Post(ROUTES.MAP.ADD_MEMORIAL_MARKER)
-  @UseGuards(JwtAuthGuard)
-  @UseInterceptors(FileFieldsInterceptor([{ name: 'photos' }]))
-  addMemorialMarker(
-    @UploadedFiles() files: { photos?: Express.Multer.File[] },
-    @Body() dto: CreateMemorialDto,
-  ) {
-    return this.mapService.addMemorial(files, dto);
-  }
-
-  @Post(ROUTES.MAP.EDIT_MEMORIAL)
-  @UseGuards(JwtAuthGuard)
-  @UseInterceptors(FileFieldsInterceptor([{ name: 'photos' }]))
-  editMemorialMarker(
-    @UploadedFiles() files: { photos?: Express.Multer.File[] },
-    @Body() dto: CreateMemorialDto,
-    @Param('id') id,
-  ) {
-    return this.mapService.editMemorialMarker(files, dto, id);
   }
 }
